txt file still only reading one line over and over again
