+++ src/com/notleh/Main.java
@@ -10,6 +10,7 @@ import com.notleh.services.InputSeperateSort;
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.util.ArrayList;
+import java.util.Objects;
 import java.util.Scanner;
 
 import static com.notleh.enums.EnumHandScores.ZER0;
@@ -27,17 +28,10 @@ public class Main {
 
     public static void main(String[] args) throws FileNotFoundException {
 
-        Scanner sc = new Scanner(new File("/Users/NLH/Desktop/poker-hands.txt"));
+        String inputString = handleInput();
 
-//        Scanner input = new Scanner(System.in);
-//        System.out.println("Input 10 card values + ENTER (eg. AH 9S 4D TD 8S 4H JS 3C TC 8D)");
-//
-//        String inputHandString = sc.nextLine();
-//        validateHand(inputHandString);
-        handleInput();
-
-        ArrayList<String> playerOneHandArr = createCardArray(inputHandString, PLAYER_ONE_BEGIN, PLAYER_ONE_END);
-        ArrayList<String> playerTwoHandArr = createCardArray(inputHandString, PLAYER_TWO_BEGIN, PLAYER_TWO_END);
+        ArrayList<String> playerOneHandArr = createCardArray(inputString, PLAYER_ONE_BEGIN, PLAYER_ONE_END);
+        ArrayList<String> playerTwoHandArr = createCardArray(inputString, PLAYER_TWO_BEGIN, PLAYER_TWO_END);
 
         Hand playerOneHand = new Hand();
         Hand playerTwoHand = new Hand();
